Implemented BinaryTreePred instead of the faulty BinarySearchTree

+++ LinearScanPred.cpp
@@ -9,7 +9,9 @@
 class LinearScanPred : public BasePred {
 public:
     /**
-     * Linearly scan through the array to find the pred
+     * Linearly scan through the array to find the pred.
+     * The array doesn't have to be sorted for this to work.
+     *
      * @param array
      * @param x
      * @return
@@ -17,7 +19,7 @@ public:
     unsigned int pred(unsigned int x) {
         unsigned int max = 0;
 
-        unsigned int size = this->array.size();
+        unsigned long size = this->array.size();
 
         for (unsigned int i = 0; i < size; ++i) {
             if (this->array.at(i) > max && this->array.at(i) <= x) {
